Module scl_gui_widgets::widgets

SCL 组件库主模块

所有的组件都在这里

Modules

label

一个标签创建模块，会使用 SCL 的主题（字体、大小等）来创建标签

Structs

Annotated

A annotated Slider or RangeSlider

AutoFocus

一个控制器，用来自动聚焦组件（例如输入框或按钮等输入组件）

Button

A button with a text label.

CreateItem

一个左侧有一个加号的可点击项目组件

DownloadModuleItem

一个左侧有图标和说明信息，右侧有副文本信息的可点击项组件

DynImage

一个可以动态改变的位图组件，可以接收 [druid::ImageBuf] 数据来改变显示的图片

FolderList

一个可以点击折叠展开的文件夹列表组件

Icon

一个图标组件，可以静态显示一个图标，也可以根据数据更新回调返回的图标键组 crate::theme::icons::IconKeyPair 显示矢量路径填充的图标

IconButton

A button with a text label.

LargeListItem

一个提供显示图标、主标签和副标签的大列表项目组件

ListItem

A button with a text label.

ListSelect

Builds a simple list selection widget, for selecting a single value out of a list.

MenuItem

一个左侧显示内容，右侧显示勾选框或文本内容的组件

NavigationControl

一个导航分页组件，提供了类似 WinUI 3 中的 Pivot 组件

OnChange

一个控制器，当数据流发生更新时触发回调

OnCmd

一个控制器，当接收到特定的 [druid::Command] 时触发回调

OnMonitor

This is a wrapper widget that attempts to ensure that the widget it wraps is fully contained in one monitor.

OnNotify

一个组件，当子组件向上抛出指定的 [druid::Notification] 时触发回调

PageSwitcher

一个分页组件，提供类似 WinUI 3 的缩放/滑动切换动画来切换所选页面

PasswordBox

自制密码输入框，可以简单的输入密码和粘贴密码

PressKey

一个控制器，当子组件被聚焦且按下指定按键时触发回调

ProgressOverlay

一个进度显示组件，用于搭配 SCL Core 的进度报告功能显示正在处理的项目进度

ProgressRing

一个仿 WinUI 3 的进度指示圈，仅支持无界进度动画

Radio

A single radio button

RadioGroup

A group of radio buttons

RangeSlider

A range slider, allowing interactive update of two numeric values .

ShowIf

一个组件，当回调返回 true 时会显示其包裹的组件，反之则隐藏且不占用布局控件

Slider

A slider, allowing interactive update of a numeric value.

SliderValueMapping

A stepped slider value mapping

StartGameButton

A button with a text label.

Store

一个中间件组件，用于拦截数据流并进行处理

ToggleSwitch

一个仿 WinUI 3 的开关组件

WindowWidget

一个用于小型窗口的仿 Windows 窗口组件，推荐创建含有此组件的窗口时将 [druid::WindowDesc] 中 [druid::WindowDesc::show_titlebar] 设置为 false，效果最佳。

Enums

KnobStyle

The shape of the slider knobs.

Constants

ADD_MAX_PROGRESS

用于给进度增加/减少最大进度值，需要传递此进度的唯一编号和其进度最大值变化值

ADD_PROGRESS

用于给进度增加/减少进度值，需要传递此进度的唯一编号和其进度变化值

BACK_PAGE_CLICKED

窗口的返回按钮被点击时抛出的 [druid::Notification]

CLEAR_BACKGROUND_IMAGE

请求清除窗口图片背景的 [druid::Command]

ENABLE_BACK_PAGE

请求启用窗口返回按钮的 [druid::Command] ，参数为是否启用/禁用

HIDE_PROGRESS

用于给进度隐藏进度值，需要传递此进度的唯一编号

NEW_PROGRESS

用于增加新进度的指令，需要传递一个独一无二的编号作为此进度的唯一编号

ON_PAGE

当分页组件跳转到指定页面时抛出的 [druid::Notification]

POP_PAGE

当分页组件即将离开指定页面时抛出的 [druid::Notification]

PUSH_PAGE

使分页组件跳转到指定页面时的 [druid::Command]

QUERY_CLOSE_WINDOW

请求关闭窗口的 [druid::Notification]

QUERY_POP_PAGE

使分页组件回到上一个或指定页面时的 [druid::Command]

REMOVE_PROGRESS

用于删除进度的指令，需要传递此进度的唯一编号

SET_BACKGROUND_IMAGE

请求设置窗口图片背景的 [druid::Command]

SET_MAX_PROGRESS

用于给进度设置最大进度值，需要传递此进度的唯一编号和其进度最大值

SET_MESSAGE

用于给进度设置主要信息，需要传递此进度的唯一编号

SET_PROGRESS

用于给进度设置进度值，需要传递此进度的唯一编号和其进度值

SET_SLIDE_PAGE_ANIMATION

使分页组件使用滑动动画而非缩放渐变动画的 [druid::Command]

SET_SUB_MESSAGE

用于给进度设置次要信息，需要传递此进度的唯一编号和其需要展示的信息

Type Definitions

ShowIfCallback

一个针对 ShowIf 触发的回调，根据其返回值决定是否显示组件