Struct scl_gui_widgets::widgets::IconButton

pub struct IconButton<T> { /* private fields */ }

A button with a text label.

Implementations

impl<T: Data> IconButton<T>

pub fn dynamic( icon_key: impl Fn(&T, &Env) -> IconKeyPair + 'static ) -> IconButton<T>

Create a new button with a text label.

Use the .on_click method to provide a closure to be called when the button is clicked.

Examples

use druid::widget::Button;

let button = Button::new("Increment").on_click(|_ctx, data: &mut u32, _env| {
    *data += 1;
});

pub fn new(icon_key: IconKeyPair) -> IconButton<T>

根据提供的图标键组对创建一个图标按钮

pub fn on_click( self, f: impl Fn(&mut EventCtx<'_, '_>, &mut T, &Env) + 'static ) -> ControllerHost<Self, Click<T>>

Provide a closure to be called when this button is clicked.

pub fn set_accent(&mut self, value: bool)

Use accent color as button color

pub fn with_accent(self, value: bool) -> Self

Builder style to use accent color as button color

pub fn set_flat(&mut self, value: bool)

Use no color as button color

pub fn with_flat(self, value: bool) -> Self

Builder style to use no color as button color